+++ pkg/worker/artifact.go
@@ -1,0 +1,6 @@
+package worker
+
+// ArtifactMeta holds metadata about an artifact created during building.
+type ArtifactMeta struct {
+	Name string
+}